+++ src/main/java/pokemon/pokedex/_common/HomeController.java
@@ -1,9 +1,10 @@
-package pokemon.pokedex;
+package pokemon.pokedex._common;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.SessionAttribute;
+import pokemon.pokedex._global.SessionConst;
 import pokemon.pokedex.user.dto.LoginResponseDTO;
 
 @Controller
