+++ src/main/java/pokemon/pokedex/user/controller/LoginController.java
@@ -9,7 +9,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
-import pokemon.pokedex.SessionConst;
+import pokemon.pokedex._global.SessionConst;
 import pokemon.pokedex.user.dto.LoginDTO;
 import pokemon.pokedex.user.dto.LoginResponseDTO;
 import pokemon.pokedex.user.exception.LoginFailedException;
