+++ src/test/java/pokemon/pokedex/HomeControllerTest.java
@@ -7,6 +7,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import pokemon.pokedex._global.SessionConst;
 import pokemon.pokedex.user.dto.LoginResponseDTO;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
