+++ src/test/java/pokemon/pokedex/user/interceptor/GuestOnlyInterceptorUnitTest.java
@@ -10,7 +10,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import pokemon.pokedex.SessionConst;
+import pokemon.pokedex._global.SessionConst;
 import pokemon.pokedex.user.controller.LoginController;
 import pokemon.pokedex.user.controller.RegisterController;
 import pokemon.pokedex.user.dto.LoginResponseDTO;
